medium filter work in progress

+++ src/common/support.h
@@ -38,7 +38,7 @@ struct CReadWrite
 template < class T >
 unsigned int convertToInt( T * _t )
 {
-	return static_cast< unsigned int >( (long long )_t );
+	return static_cast< unsigned long long >( (long long )_t );
 }
 
 }
